requirements.txt
torch~=2.8.0
chess~=1.11.2
msgpack-python~=0.5.6
numpy~=2.3.2
scipy~=1.16.1
pip~=24.3.1
filelock~=3.18.0
Flask~=3.1.1
flask-cors~=6.0.1
Flask-Limiter~=3.12